+++ src/main/java/org/sk/skMinecraft/CommandFactory.java
@@ -24,6 +24,7 @@ public class CommandFactory {
 
         this.commands.put("spawnEntity", SpawnEntity::new);
         this.commands.put("editEntity", EditEntity::new);
+        this.commands.put("getEntity", GetEntity::new);
 
         this.commands.put("addInv", AddInventory::new);
         this.commands.put("getInv", GetInventory::new);